Startup-Chatbot/requirements.txt
# Core
boto3>=1.34.0
botocore>=1.34.0

# LangChain ecosystem
langchain>=0.2.2
langchain-community>=0.2.2
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.3

# Vector store + embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document loaders
pypdf>=4.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.3
docx2txt>=0.8
unstructured>=0.14.0

# Web + search adapters
requests>=2.31.0
tenacity>=8.2.2

# JSON parsing/repair
json5>=0.9.14

# Utils
tqdm>=4.66.0

# Optional: for plotting peer positioning charts later
matplotlib>=3.8.0


# --- Core ---
python>=3.10
google-cloud-aiplatform>=1.71.1
google-generativeai>=0.8.3
google-cloud-documentai>=2.29.0

# --- LangChain / Vector DB ---
langchain>=0.2.12
langchain-community>=0.2.11
faiss-cpu>=1.8.0
tiktoken>=0.7.0

# --- PDF / OCR ---
PyMuPDF>=1.24.9
pdfplumber>=0.11.0

# --- Data Processing ---
pandas>=2.2.2
numpy>=1.26.4

# --- JSON handling ---
json5>=0.9.25

# --- Web requests / scraping ---
httpx>=0.27.0
beautifulsoup4>=4.12.3

# --- Utils ---
python-dotenv>=1.0.1
tenacity>=8.4.1